Add argument for bash file main.sh Fixes #1
