+++ .gitignore
@@ -122,3 +122,8 @@ ENV/
 
 # Checkpoint model
 *.ckpt*
+ckpts/
+setup.cfg
+training/
+datasets/
+.vscode/
