+++ ckpts/checkpoint
@@ -1,2 +1,6 @@
-model_checkpoint_path: "model.ckpt-0"
-all_model_checkpoint_paths: "model.ckpt-0"
+model_checkpoint_path: "model.ckpt-10000"
+all_model_checkpoint_paths: "model.ckpt-4604"
+all_model_checkpoint_paths: "model.ckpt-6151"
+all_model_checkpoint_paths: "model.ckpt-7698"
+all_model_checkpoint_paths: "model.ckpt-9241"
+all_model_checkpoint_paths: "model.ckpt-10000"
