+++ datasets/test_labels.csv
@@ -1,56 +1,55 @@
 filename,width,height,class,xmin,ymin,xmax,ymax
-raccoon-10.jpg,450,495,raccoon,130,2,446,488
-raccoon-101.jpg,640,426,raccoon,86,53,400,356
-raccoon-11.jpg,660,432,raccoon,3,1,461,431
-raccoon-12.jpg,259,194,raccoon,28,21,126,181
-raccoon-12.jpg,259,194,raccoon,85,33,235,193
+raccoon-102.jpg,259,194,raccoon,1,1,118,152
+raccoon-105.jpg,720,960,raccoon,250,49,714,869
+raccoon-108.jpg,604,481,raccoon,99,53,402,464
+raccoon-116.jpg,660,432,raccoon,3,1,436,430
 raccoon-120.jpg,660,371,raccoon,129,12,510,331
-raccoon-127.jpg,253,199,raccoon,125,59,212,165
-raccoon-130.jpg,640,426,raccoon,223,62,497,307
-raccoon-130.jpg,640,426,raccoon,453,41,640,423
-raccoon-132.jpg,259,194,raccoon,6,2,240,131
-raccoon-133.jpg,490,640,raccoon,8,6,476,631
-raccoon-134.jpg,225,225,raccoon,125,87,194,169
-raccoon-140.jpg,204,247,raccoon,6,17,202,231
+raccoon-13.jpg,660,495,raccoon,55,28,393,313
+raccoon-135.jpg,640,426,raccoon,99,8,605,404
+raccoon-137.jpg,320,240,raccoon,71,8,304,233
+raccoon-14.jpg,900,484,raccoon,163,81,546,438
 raccoon-141.jpg,249,202,raccoon,1,1,154,176
-raccoon-142.jpg,1024,768,raccoon,171,162,811,740
-raccoon-146.jpg,275,183,raccoon,4,4,271,180
+raccoon-143.jpg,259,194,raccoon,17,29,238,162
+raccoon-144.jpg,570,390,raccoon,117,42,387,390
+raccoon-145.jpg,600,450,raccoon,3,36,345,450
+raccoon-145.jpg,600,450,raccoon,260,41,569,449
 raccoon-150.jpg,275,183,raccoon,80,62,187,169
 raccoon-151.jpg,225,225,raccoon,42,94,108,224
-raccoon-153.jpg,700,700,raccoon,10,1,612,700
-raccoon-163.jpg,248,203,raccoon,6,7,240,157
+raccoon-158.jpg,275,183,raccoon,23,23,262,177
+raccoon-159.jpg,226,223,raccoon,14,11,223,221
+raccoon-160.jpg,256,197,raccoon,7,42,162,197
+raccoon-161.jpg,500,347,raccoon,209,73,385,186
+raccoon-162.jpg,259,194,raccoon,45,34,161,184
+raccoon-167.jpg,259,195,raccoon,1,5,175,195
 raccoon-172.jpg,615,346,raccoon,183,53,399,302
-raccoon-174.jpg,960,639,raccoon,125,43,588,527
 raccoon-175.jpg,634,381,raccoon,69,89,354,378
-raccoon-177.jpg,276,183,raccoon,8,18,157,178
-raccoon-177.jpg,276,183,raccoon,146,13,263,146
-raccoon-179.jpg,600,450,raccoon,1,176,270,427
-raccoon-181.jpg,750,422,raccoon,100,1,420,411
 raccoon-183.jpg,2000,1333,raccoon,358,21,1354,1119
-raccoon-188.jpg,460,379,raccoon,26,71,366,334
-raccoon-191.jpg,634,445,raccoon,100,89,478,331
-raccoon-198.jpg,259,194,raccoon,57,21,158,184
-raccoon-198.jpg,259,194,raccoon,112,32,199,158
-raccoon-199.jpg,640,428,raccoon,28,64,530,402
-raccoon-22.jpg,640,360,raccoon,252,76,466,335
+raccoon-184.jpg,640,640,raccoon,81,77,567,617
+raccoon-19.jpg,259,194,raccoon,87,8,182,89
+raccoon-21.jpg,290,174,raccoon,59,2,216,171
 raccoon-23.jpg,259,194,raccoon,108,1,258,194
-raccoon-28.jpg,602,452,raccoon,93,80,601,452
-raccoon-32.jpg,625,415,raccoon,88,92,473,328
-raccoon-35.jpg,275,183,raccoon,1,1,164,183
-raccoon-36.jpg,640,428,raccoon,113,27,468,428
-raccoon-45.jpg,620,372,raccoon,140,6,454,370
-raccoon-48.jpg,261,193,raccoon,43,28,240,176
-raccoon-54.jpg,602,339,raccoon,78,5,517,333
-raccoon-55.jpg,634,417,raccoon,6,49,250,320
-raccoon-55.jpg,634,417,raccoon,274,27,563,410
+raccoon-26.jpg,306,374,raccoon,114,5,306,337
+raccoon-38.jpg,259,194,raccoon,7,17,257,180
+raccoon-39.jpg,250,172,raccoon,54,12,250,166
+raccoon-40.jpg,480,360,raccoon,164,53,349,275
+raccoon-41.jpg,700,500,raccoon,211,78,530,468
+raccoon-47.jpg,262,193,raccoon,34,4,233,193
+raccoon-5.jpg,270,187,raccoon,3,3,260,179
+raccoon-51.jpg,800,599,raccoon,315,105,772,540
+raccoon-52.jpg,800,533,raccoon,105,10,502,501
+raccoon-53.jpg,259,194,raccoon,71,45,197,171
+raccoon-57.jpg,640,425,raccoon,82,6,638,423
 raccoon-58.jpg,224,225,raccoon,2,1,199,221
-raccoon-6.jpg,480,360,raccoon,1,44,307,316
-raccoon-62.jpg,640,407,raccoon,73,19,632,407
-raccoon-70.jpg,500,375,raccoon,60,4,421,369
+raccoon-61.jpg,274,184,raccoon,94,63,195,148
+raccoon-61.jpg,274,184,raccoon,142,39,213,108
+raccoon-63.jpg,600,400,raccoon,74,107,280,290
+raccoon-63.jpg,600,400,raccoon,227,93,403,298
+raccoon-65.jpg,480,360,raccoon,123,27,338,284
+raccoon-72.jpg,560,420,raccoon,219,195,446,375
+raccoon-72.jpg,560,420,raccoon,98,34,284,336
+raccoon-74.jpg,800,533,raccoon,141,6,472,505
 raccoon-75.jpg,640,640,raccoon,1,1,640,459
-raccoon-76.jpg,225,225,raccoon,14,1,212,132
-raccoon-79.jpg,640,425,raccoon,120,1,568,425
-raccoon-83.jpg,660,371,raccoon,104,3,509,369
 raccoon-84.jpg,303,166,raccoon,31,6,197,163
-raccoon-89.jpg,259,194,raccoon,18,6,225,176
-raccoon-98.jpg,480,360,raccoon,108,31,351,308
+raccoon-91.jpg,236,314,raccoon,22,14,216,308
+raccoon-95.jpg,320,400,raccoon,50,45,272,289
+raccoon-96.jpg,230,219,raccoon,28,25,203,175
