+++ requirements.txt
@@ -7,6 +7,7 @@ matplotlib
 tf_slim
 pycocotools
 pandas
-numpy
+numpy==1.17.4
 requests
 tqdm
+pyfiglet
