--- test.py
@@ -1,5 +1,0 @@
-import tensorflow as tf
-
-tf.test.is_gpu_available(
-    cuda_only=False, min_cuda_compute_capability=None
-)
